進度: 正在製作WordActivity的生命週期規劃

+++ app/src/main/java/com/bu/selfstudy/logic/dao/WordDao.kt
@@ -12,7 +12,7 @@ interface WordDao {
     fun findWordsByName(name:String): List<Word>
 
     @Query("SELECT * FROM Word WHERE bookId=:bookId AND name LIKE :name")
-    fun findWordsByNameAndBookId(name:String, bookId:Int): List<Word>
+    fun findWordsByNameAndBookId(name:String, bookId:Long): List<Word>
 
     @Update
     fun updateWords(vararg words: Word):Int
